Destroying app logic moved to go client, moved UpdateEnv to app.go
