+++ helpers/app.go
@@ -24,9 +24,6 @@ func DeployApp(client *deploy.DeployAPIV1, token runtime.ClientAuthInfoWriter, a
 	app_req := models.AppRequest21{Type: &req_type, Env: env, ContainerCount: 1, ContainerSize: 1024}
 	app_params := operations.NewPostAppsAppIDOperationsParams().WithAppID(app_id).WithAppRequest(&app_req)
 	_, err := client.Operations.PostAppsAppIDOperations(app_params, token)
-	if err != nil {
-		return err
-	}
 	return err
 }
 
@@ -49,4 +46,33 @@ func GetApp(client *deploy.DeployAPIV1, token runtime.ClientAuthInfoWriter, app_
 		}
 	}
 	return false, err
+}
+
+// Updates the `env` based on changes made in the config file
+func UpdateEnv(env map[string]interface{}, client *deploy.DeployAPIV1, app_id int64, bearerTokenAuth runtime.ClientAuthInfoWriter) error {
+	app_req := models.AppRequest21{}
+	if _, ok := env["APTIBLE_DOCKER_IMAGE"]; ok {
+		// Deploying app
+		req_type := "deploy"
+		app_req = models.AppRequest21{Type: &req_type, Env: env, ContainerCount: 1, ContainerSize: 1024}
+	} else {
+		// Configuring app
+		req_type := "configure"
+		app_req = models.AppRequest21{Type: &req_type, Env: env}
+	}
+
+	app_params := operations.NewPostAppsAppIDOperationsParams().WithAppID(app_id).WithAppRequest(&app_req)
+	_, err := client.Operations.PostAppsAppIDOperations(app_params, bearerTokenAuth)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func DestroyApp(client *deploy.DeployAPIV1, token runtime.ClientAuthInfoWriter, app_id int64) error {
+	req_type := "deprovision"
+	app_req := models.AppRequest21{Type: &req_type}
+	app_params := operations.NewPostAppsAppIDOperationsParams().WithAppID(app_id).WithAppRequest(&app_req)
+	_, err := client.Operations.PostAppsAppIDOperations(app_params, token)
+	return err
 }
